+++ internal/httpapi/assets/_app/immutable/entry/start.nmm08JTP.js
@@ -1,0 +1,1 @@
+import{l as o,a as r}from"../chunks/CwfF3OcA.js";export{o as load_css,r as start};